goal/goal-static/requirements.txt
customtkinter
ctkmessagebox
python-dotenv
pymongo
pyinstaller
keyboard
cryptography